+++ src/gfunction/gfunction.go
@@ -14,6 +14,12 @@ import (
 	"jacobin/types"
 )
 
+// Radix boundaries:
+var MinRadix int64 = 2
+var MaxRadix int64 = 36
+var MaxIntValue int64 = 2147483647
+var MinIntValue int64 = -2147483648
+
 // GMeth is the entry in the MTable for Go functions. See MTable comments for details.
 //   - ParamSlots - the number of user parameters in a G function. E.g. For atan2, this would be 2.
 //   - GFunction - a go function. All go functions accept a possibly empty slice of interface{} and
@@ -44,12 +50,19 @@ func getGErrBlk(exceptionType int, errMsg string) *GErrBlk {
 // they make available.
 func MTableLoadNatives(MTable *classloader.MT) {
 
-	loadlib(MTable, Load_Io_PrintStream())         // load the java.io.prinstream golang functions
-	loadlib(MTable, Load_Io_Console())             // load the java.io.Console golang functions
-	loadlib(MTable, Load_Lang_Class())             // load the java.lang.Class golang functions
-	loadlib(MTable, Load_Lang_Math())              // load the java.lang.Math golang functions
-	loadlib(MTable, Load_Lang_Object())            // load the java.lang.Class golang functions
-	loadlib(MTable, Load_Misc_Unsafe())            // load the jdk.internal/misc/Unsafe functions
+	loadlib(MTable, Load_Io_Console())     // load the java.io.Console golang functions
+	loadlib(MTable, Load_Io_PrintStream()) // load the java.io.prinstream golang functions
+	loadlib(MTable, Load_Lang_Boolean())
+	loadlib(MTable, Load_Lang_Byte())
+	loadlib(MTable, Load_Lang_Character())
+	loadlib(MTable, Load_Lang_Class()) // load the java.lang.Class golang functions
+	loadlib(MTable, Load_Lang_Double())
+	loadlib(MTable, Load_Lang_Float())
+	loadlib(MTable, Load_Lang_Integer())
+	loadlib(MTable, Load_Lang_Long())
+	loadlib(MTable, Load_Lang_Math())   // load the java.lang.Math & StrictMath golang functions
+	loadlib(MTable, Load_Lang_Object()) // load the java.lang.Class golang functions
+	loadlib(MTable, Load_Lang_Short())
 	loadlib(MTable, Load_Lang_String())            // load the java.lang.String golang functions
 	loadlib(MTable, Load_Lang_StringBuilder())     // load the java.lang.StringBuilder golang functions
 	loadlib(MTable, Load_Lang_System())            // load the java.lang.System golang functions
@@ -57,9 +70,9 @@ func MTableLoadNatives(MTable *classloader.MT) {
 	loadlib(MTable, Load_Lang_Thread())            // load the java.lang.Thread golang functions
 	loadlib(MTable, Load_Lang_Throwable())         // load the java.lang.Throwable golang functions (errors & exceptions)
 	loadlib(MTable, Load_Lang_UTF16())             // load the java.lang.UTF16 golang functions
+	loadlib(MTable, Load_Misc_Unsafe())            // load the jdk.internal/misc/Unsafe functions
 	loadlib(MTable, Load_Util_HashMap())           // load the java.util.HashMap golang functions
 	loadlib(MTable, Load_Util_Locale())            // load the java.util.Locale golang functions
-	loadlib(MTable, Load_Primitives())             // load the Java primitives golang functions
 }
 
 func loadlib(tbl *classloader.MT, libMeths map[string]GMeth) {
