+++ src/gfunction/javaLangFloat.go
@@ -1,0 +1,18 @@
+/*
+ * Jacobin VM - A Java virtual machine
+ * Copyright (c) 2023 by  the Jacobin authors. Consult jacobin.org.
+ * Licensed under Mozilla Public License 2.0 (MPL 2.0) All rights reserved.
+ */
+
+package gfunction
+
+func Load_Lang_Float() map[string]GMeth {
+
+	MethodSignatures["java/lang/Float.<clinit>()V"] =
+		GMeth{
+			ParamSlots: 0,
+			GFunction:  justReturn,
+		}
+
+	return MethodSignatures
+}
